+++ packages/plugin-loans-api/src/configs.ts
@@ -6,6 +6,7 @@ import documents from './documents';
 import forms from './forms';
 import imports from './imports';
 import exporter from './exporter';
+import logs from './logUtils';
 import * as permissions from './permissions';
 import payment from './payment';
 import { checkContractScheduleAnd } from './cronjobs/contractCronJobs';
@@ -46,6 +47,7 @@ export default {
     graphqlPubsub = options.pubsubClient;
   },
   meta: {
+    logs: { consumers: logs },
     cronjobs: {
       handleDailyJob: checkContractScheduleAnd
     },
