+++ packages/plugin-loans-api/src/graphql/resolvers/mutations/invoices.ts
@@ -1,16 +1,10 @@
-import { gatherDescriptions } from '../../../utils';
-import {
-  checkPermission,
-  putCreateLog,
-  putDeleteLog,
-  putUpdateLog
-} from '@erxes/api-utils/src';
+import { checkPermission } from '@erxes/api-utils/src';
 import { IContext } from '../../../connectionResolver';
-import messageBroker from '../../../messageBroker';
 import {
   IInvoice,
   IInvoiceDocument
 } from '../../../models/definitions/invoices';
+import { createLog, deleteLog, updateLog } from '../../../logUtils';
 
 const invoiceMutations = {
   loanInvoicesAdd: async (
@@ -31,14 +25,7 @@ const invoiceMutations = {
       extraParams: { models }
     };
 
-    const descriptions = gatherDescriptions(logData);
-
-    await putCreateLog(
-      subdomain,
-      messageBroker(),
-      { ...descriptions, ...logData },
-      user
-    );
+    await createLog(subdomain, user, logData);
 
     return invoice;
   },
@@ -66,14 +53,7 @@ const invoiceMutations = {
       extraParams: { models }
     };
 
-    const descriptions = gatherDescriptions(logData);
-
-    await putUpdateLog(
-      subdomain,
-      messageBroker(),
-      { ...descriptions, ...logData },
-      user
-    );
+    await updateLog(subdomain, user, logData);
 
     return updated;
   },
@@ -100,13 +80,7 @@ const invoiceMutations = {
         object: invoice,
         extraParams: { models }
       };
-      const descriptions = gatherDescriptions(logData);
-      await putDeleteLog(
-        subdomain,
-        messageBroker(),
-        { ...logData, ...descriptions },
-        user
-      );
+      await deleteLog(subdomain, user, logData);
     }
 
     return invoiceIds;
