+++ packages/plugin-loans-api/src/graphql/resolvers/mutations/transactions.ts
@@ -1,16 +1,11 @@
-import { gatherDescriptions } from '../../../utils';
-import {
-  checkPermission,
-  putCreateLog,
-  putDeleteLog,
-  putUpdateLog
-} from '@erxes/api-utils/src';
+import { checkPermission } from '@erxes/api-utils/src';
 import { IContext } from '../../../connectionResolver';
-import messageBroker, { sendMessageBroker } from '../../../messageBroker';
+import { sendMessageBroker } from '../../../messageBroker';
 import {
   ITransaction,
   ITransactionDocument
 } from '../../../models/definitions/transactions';
+import { createLog, deleteLog, updateLog } from '../../../logUtils';
 
 const transactionMutations = {
   transactionsAdd: async (
@@ -30,17 +25,7 @@ const transactionMutations = {
       extraParams: { models }
     };
 
-    const descriptions = gatherDescriptions(logData);
-
-    await putCreateLog(
-      subdomain,
-      messageBroker(),
-      {
-        ...logData,
-        ...descriptions
-      },
-      user
-    );
+    await createLog(subdomain, user, logData);
 
     return transaction;
   },
@@ -72,17 +57,7 @@ const transactionMutations = {
       extraParams: { models }
     };
 
-    const descriptions = gatherDescriptions(logData);
-
-    await putUpdateLog(
-      subdomain,
-      messageBroker(),
-      {
-        ...logData,
-        ...descriptions
-      },
-      user
-    );
+    await updateLog(subdomain, user, logData);
 
     return updated;
   },
@@ -110,17 +85,7 @@ const transactionMutations = {
       extraParams: { models }
     };
 
-    const descriptions = gatherDescriptions(logData);
-
-    await putUpdateLog(
-      subdomain,
-      messageBroker(),
-      {
-        ...logData,
-        ...descriptions
-      },
-      user
-    );
+    await updateLog(subdomain, user, logData);
 
     return updated;
   },
@@ -149,7 +114,6 @@ const transactionMutations = {
         extraParams: { models }
       };
 
-      const descriptions = gatherDescriptions(logData);
       if (!!transaction.ebarimt && transaction.isManual)
         await sendMessageBroker(
           {
@@ -163,13 +127,7 @@ const transactionMutations = {
           'ebarimt'
         );
 
-      await putDeleteLog(
-        subdomain,
-        messageBroker(),
-
-        { ...logData, ...descriptions },
-        user
-      );
+      await deleteLog(subdomain, user, logData);
     }
 
     return transactionIds;
